# backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Share Backend API",
    "description": "Share Backend API",
    "version": "v0.1.0"
  },
  "servers": [
    {
      "url": "https://hub.maayuan.fun:16666",
      "description": "Public HTTPS endpoint"
    }
  ],
...
  "paths": {
...
    "/user/open-api/token": {
      "post": {
        "tags": [
          "OpenAPI Token"
        ],
        "summary": "生成第三方 API Token",
        "description": "生成第三方 API Token(需登录)",
        "operationId": "generate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenApiTokenGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultOpenApiTokenCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "404": {
            "description": "Inventory account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "429": {
            "description": "Token limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
...
    "/user/open-api/tokens": {
      "get": {
        "tags": [
          "OpenAPI Token"
        ],
        "summary": "列举当前用户的 token",
        "description": "列举当前用户的 token(需登录)",
        "operationId": "tokens",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListOpenApiTokenListItemDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
...
    "/user/open-api/tokens/{tokenId}": {
      "delete": {
        "tags": [
          "OpenAPI Token"
        ],
        "summary": "删除第三方 API Token",
        "description": "删除第三方 API Token(需登录)",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultUnit"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResultUnit": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status_code"
        ]
      },
...
      "OpenApiTokenGenerateRequest": {
        "properties": {
          "account_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "inventory:read",
                "inventory:write",
                "inventory:export"
              ]
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "remark": {
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "scopes"
        ]
      },
      "ApiResult": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "status_code"
        ]
      },
      "ApiResultOpenApiTokenCreatedResponse": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/OpenApiTokenCreatedResponse"
          }
        },
        "required": [
          "status_code"
        ]
      },
      "OpenApiTokenCreatedResponse": {
        "properties": {
          "token_id": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "account_id": {
            "type": "string"
          },
          "account_name": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "account_id",
          "account_name",
          "created_at",
          "scopes",
          "token",
          "token_id"
        ]
      },
...
      "ApiResultListOpenApiTokenListItemDto": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenApiTokenListItemDto"
            }
          }
        },
        "required": [
          "status_code"
        ]
      },
      "OpenApiTokenListItemDto": {
        "properties": {
          "token_id": {
            "type": "string"
          },
          "account_id": {
            "type": "string"
          },
          "account_name": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "account_id",
          "account_name",
          "created_at",
          "scopes",
          "token_id"
        ]
      },
...
    },
    "securitySchemes": {
      "Jwt": {
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer"
      },
      "OpenApiToken": {
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "API token"
      }
    }
  }
}